DataPipelineMonitoring/requirements.txt
django

mongoengine

django-cors-headers

python-crontab

pymongo

pytest

coverage

requests

pandas
